requirements.txt
aiogram>=3.15.0
aiohttp>=3.9.5
chromadb==0.5.5
pypdf==4.3.1
duckduckgo-search==6.2.4
python-dotenv==1.0.1
aiosqlite==0.20.0
reportlab>=4.0.0
python-docx==1.1.2
python-pptx==1.0.2
openai==1.43.0
pydantic>=2.9.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9